Improved JWT token handling
