+++ Sparc.Blossom/Sparc.Blossom.csproj
@@ -3,7 +3,7 @@
 		<TargetFrameworks>net7.0</TargetFrameworks>
 		<!--<TargetFrameworks>net7.0;net7.0-ios;net7.0-android;net7.0-maccatalyst</TargetFrameworks>-->
 		<!--<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows'))">$(TargetFrameworks);net7.0-windows10.0.19041.0</TargetFrameworks>-->
-		<Version>7.0.26</Version>
+		<Version>7.0.28</Version>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<UseMaui>true</UseMaui>
 		<SingleProject>true</SingleProject>
@@ -79,6 +79,7 @@
 		<PackageReference Include="Microsoft.Authentication.WebAssembly.Msal" Version="7.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Http.Polly" Version="7.0.0" />
 		<PackageReference Include="Blazored.LocalStorage" Version="4.2.0" />
+		<PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.25.0" />
 	</ItemGroup>
 
 	<ItemGroup Condition="$(TargetFramework.Contains('android'))">
